requirements.txt
audioread==2.1.6
Click==7.0
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
Flask-SocketIO==3.0.2
Flask-WTF==0.14.2
gunicorn==19.7.1
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.0
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.25.0
MarkupSafe==1.1.0
matplotlib==3.0.2
numba==0.40.1
numpy==1.15.4
passlib==1.7.1
pymongo==3.5.1
pyparsing==2.3.0
python-dateutil==2.7.5
python-engineio==2.3.2
python-socketio==2.0.0
redis==3.0.1
resampy==0.2.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
Werkzeug==0.14.1
WTForms==2.2.1
aiohttp
python-socketio
eventlet
flask-cors

